New Migrations, Created Identity Scaffolding, Updated Register, Updated Home Page
Deleted existing Migrations and Added a new Migrations and updated the database

Created Identity Scaffolding and included all options.

Updated Registry. Added the First Name and Last Name on the field as a requirement to register.

Updated Home Page. When logged in, it now displays the complete name of the User